Rework articulation and note duration
* Move setting of initial tempo into score
* Replace NoteOptions with Articulation and add staccato and legato
* Drop NoteBehavior and slide syntax; add field for `sustained`
* Drop NoteOff event; store end time in NoteEvent.

Computing end time in score allows the timeline creation logic to
handle adjusting times for articulation and correctly handling
sustained notes so the logic doesn't have to be duplicated across
generators. This also fixes a bug in handling of repeated, sustained
notes.

Dropped Slide because it would be very hard to implement properly with
MIDI, and it's not clear that the syntax should be conflated with tie
at the end of a note.